backend/pyproject.toml
[tool.poetry]
name = "sticky-brownie"
version = "0.1.0"
description = "온라인 여행사 서비스"
authors = ["Spakky <sejong418@icloud.com>"]
packages = [{ include = "*", from = "src" }]
readme = "README.md"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.11"
spakky-fastapi = "^0.7.16"
uvicorn = "^0.27.1"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.27" }
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"
psycopg = { extras = ["binary"], version = "^3.1.18" }
aiofiles = "^23.2.1"


[tool.poetry.group.dev.dependencies]
pylint = "^3.0.2"
black = "^23.11.0"
isort = "^5.12.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
httpx = "^0.26.0"
alembic = "^1.13.1"

[tool.poetry.group.ci.dependencies]
pylint = "^3.0.2"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
alembic = "^1.13.1"

[tool.black]
line-length = 90
target-version = ["py310"]
include = '\.pyi?$'

[tool.isort]
py_version = 311
src_paths = ["src", "tests"]
skip = [".gitignore", ".dockerignore"]
extend_skip = [".md", ".json", ".txt"]
profile = "black"
combine_as_imports = true
remove_redundant_aliases = true
length_sort_sections = ["future", "stdlib"]

[tool.pylint.MASTER]
ignore-patterns = "test_.*?py"

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 90
disable = """
    no-member,
    arguments-differ,
    invalid-field-call,
    too-many-arguments,
    missing-module-docstring,
    missing-class-docstring,
    missing-function-docstring,
    too-many-instance-attributes,
    too-few-public-methods,
    unnecessary-ellipsis,
    redefined-builtin,
    too-many-locals,
    wrong-import-order,
"""

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = "tests"
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = """
    --cov
    --cov-report=term
    --cov-report=xml
    --no-cov-on-fail
    --strict-markers
    -p no:warnings
    -vv
"""

[tool.coverage.run]
omit = ["tests/*"]
traceback = "short"
branch = true

[tool.coverage.report]
show_missing = true
precision = 2
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "@(abc\\.)?abstractmethod",
    "@(typing\\.)?overload",
]
